Adding make request related changes 2

+++ src/main/java/edu/esoft/sdp/pickAndGoBackend/service/BranchService.java
@@ -14,9 +14,9 @@ public interface BranchService {
 
     List<Branch> getAllBranches(Integer branchID);
 
-    Branch getBranchById(Integer id) throws Exception;
+    Branch getBranchById(Long id) throws Exception;
 
-    String deleteBranch(Integer id);
+    String deleteBranch(Long id);
 
     Branch updateBranch(Branch branch);
